+++ src/examples/gltf_test.rs
@@ -24,36 +24,36 @@ pub fn main() {
     // );
 
     engine.add_gltf(Transform {
-            position: [30.0, 0.0, 0.0],
+            position: [5.0, 100.0, 5.0],
             scale: [10.0, 10.0, 10.0],
             ..Default::default()
         },
         &Material::standard()
-            .color([0.1, 1.0, 0.1])
+            .color([1.0,1.0,1.0])
             .shader(ShaderType::Pbr)
             .build(),
         &Physics::default()
-            .compute_shader(ComputeShaderType::FullPhysics)
+            .compute_shader(ComputeShaderType::Test)
             .mass(10.0)
             .gravity_scale(1.0)
             .collision_type(CollisionType::Sphere),
         "./testModels/1kRustingSphere.gltf"
     );
 
-    for i in 0..1 {
-        for j in 0..1 {
-            for k in 0..1 {
+    for i in 0..10 {
+        for j in 0..10 {
+            for k in 0..10 {
                 engine.add_gltf(Transform {
-                    position: [0.0 + (5.0*i as f32), 0.0 + (5.0*j as f32), 0.0 + (5.0*k as f32)],
+                    position: [0.0 + (1.0*i as f32), 0.0 + (1.0*j as f32), 0.0 + (1.0*k as f32)],
                     scale: [1.0, 1.0, 1.0],
                     ..Default::default()
                 },
                 &Material::standard()
-                    .color([0.1, 1.0, 0.1])
+                    .color([1.0, 1.0, 1.0])
                     .shader(ShaderType::Pbr)
                     .build(),
                 &Physics::default()
-                    .compute_shader(ComputeShaderType::FullPhysics)
+                    .compute_shader(ComputeShaderType::Test)
                     .mass(10.0)
                     .gravity_scale(1.0)
                     .collision_type(CollisionType::Box),